Spectral scan should not reset the FPGA
The spectral scan should not reset the FPGA and should not need to set the NOTCH FREQ as it has already be set in the packet forwarder setup.

Calling lgw_connect to reset the FPGA and reconfigure will change the register page in the Sx1301 causing a timersync issue in the packet forwarder. Downlinks cannot be scheduled after spectral scan has run because the timestamp from the get_trigcnt is no longer valid.

THe LGW_TIMESTAMP register is on page 2 but the Sx1301 was reset to page 0. The packet forwarder is not aware of the page change and cannot recover.