Allow #define configuration of SPI parameters
Enclose preprocessor define of SPI_SPEED and SPI_DEV_PATH in #ifndef
so that these values can optinally be passed by compiler / makefile
defines at build time for a specific platform or architecture.

Allow CFLAGS to be passed in to Makefile (e.g. to allow preprocessor
defines to be set)

Signed-off-by: Scott Wagner <scott.wagner@promethean-design.com>