Add optional command line argument for TTY device
The serial port "/dev/ttyAMA0" is not necessarily the correct TTY
for the GPS device.  Hence test_loragw_gps may attempt to access
the wrong serial port for the GPS if it is hard-wired to this
device file.  This change allows optionally specifying a TTY device
for test_loragw_gps.

Signed-off-by: Scott Wagner <scott.wagner@promethean-design.com>